added separate bin for seeders
