+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod handlers;
+pub mod scrapers;
+pub mod models;
+pub mod utils;
+pub mod seeders;
+pub mod entity;
